requirements.txt
Django~=3.2.6
whitenoise==4.1.2
gunicorn==19.9.0
raven==6.10.0
psycopg2==2.9.1
python-decouple==3.4
django-utils-six==2.0
djangorestframework==3.12.4
pillow==8.0.1
